LIB-109 - add control to avoid Typed property error from PHP

+++ src/Order/Worker.php
@@ -105,7 +105,7 @@ trait Worker
 
     public function __destruct()
     {
-        if ($this->queue) {
+        if (!empty($this->queue) && $this->queue instanceof AMQPQueue) {
             $this->queue->delete();
         }
         $this->disconnect();
